--- docs/src/pages/docs/api/combine.en.mdx
@@ -1,71 +1,0 @@
----
-title: combine
----
-
-import { Sandpack } from '@/components/Sandpack';
-
-# combine 
-
-## combineCharacter
-
-Given inputs of an choseong, jungseong, and jongseong return a single Korean character
-
-```typescript
-function combineCharacter(choseong: string, jungseong: string, jongseong?: string): string;
-```
-
-### Examples
-
-```typescript
-combineCharacter('ㄱ', 'ㅏ', 'ㅂㅅ') // '값'
-combineCharacter('ㅌ', 'ㅗ') // '토'
-```
-
-### Demo
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { combineCharacter } from 'es-hangul';
-
-console.log(combineCharacter('ㄱ', 'ㅏ', 'ㅂㅅ')) 
-console.log(combineCharacter('ㅌ', 'ㅗ')) 
-
-```
-
-</Sandpack>
-
-
-## combineVowels
-
-Given two vowel inputs, combine them to create a diphthong. If the vowels cannot be combined according to the correct Korean rules, simply join them together
-
-```typescript
-function combineVowels(vowel1: string, vowel2: string): string 
-```
-
-### Examples
-
-```typescript
-combineVowels('ㅗ', 'ㅏ') // 'ㅘ'
-combineVowels('ㅗ', 'ㅐ') // 'ㅙ'
-combineVowels('ㅗ', 'ㅛ') // 'ㅗㅛ'
-```
-
-### Demo
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { combineVowels } from 'es-hangul';
-
-console.log(combineVowels('ㅗ', 'ㅏ'))
-console.log(combineVowels('ㅗ', 'ㅐ'))
-console.log(combineVowels('ㅗ', 'ㅛ'))
-
-```
-</Sandpack>
