--- docs/src/pages/docs/api/combine.ko.mdx
@@ -1,71 +1,0 @@
----
-title: combine
----
-
-import { Sandpack } from '@/components/Sandpack';
-
-# combine 
-
-## combineCharacter
-
-인자로 받은 초성, 중성, 종성을 조합해 한글 1자를 반환합니다.
-
-```typescript
-function combineCharacter(choseong: string, jungseong: string, jongseong?: string): string;
-```
-
-### Examples
-
-```typescript
-combineCharacter('ㄱ', 'ㅏ', 'ㅂㅅ') // '값'
-combineCharacter('ㅌ', 'ㅗ') // '토'
-```
-
-### 사용해보기
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { combineCharacter } from 'es-hangul';
-
-console.log(combineCharacter('ㄱ', 'ㅏ', 'ㅂㅅ')) 
-console.log(combineCharacter('ㅌ', 'ㅗ')) 
-
-```
-
-</Sandpack>
-
-
-## combineVowels
-
-인자로 두 개의 모음을 받아 합성하여 겹모음을 생성합니다. 만약 올바른 한글 규칙으로 합성할 수 없는 모음들이라면 단순 Join합니다.
-
-```typescript
-function combineVowels(vowel1: string, vowel2: string): string 
-```
-
-### Examples
-
-```typescript
-combineVowels('ㅗ', 'ㅏ') // 'ㅘ'
-combineVowels('ㅗ', 'ㅐ') // 'ㅙ'
-combineVowels('ㅗ', 'ㅛ') // 'ㅗㅛ'
-```
-
-### 사용해보기
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { combineVowels } from 'es-hangul';
-
-console.log(combineVowels('ㅗ', 'ㅏ'))
-console.log(combineVowels('ㅗ', 'ㅐ'))
-console.log(combineVowels('ㅗ', 'ㅛ'))
-
-```
-</Sandpack>
