+++ docs/src/pages/docs/api/combineCharacter.ko.mdx
@@ -1,0 +1,35 @@
+---
+title: combineCharacter
+---
+
+import { Sandpack } from '@/components/Sandpack';
+
+# combineCharacter
+
+인자로 받은 초성, 중성, 종성을 조합해 한글 1자를 반환합니다.
+
+```typescript
+function combineCharacter(choseong: string, jungseong: string, jongseong?: string): string;
+```
+
+## Examples
+
+```typescript
+combineCharacter('ㄱ', 'ㅏ', 'ㅂㅅ'); // '값'
+combineCharacter('ㅌ', 'ㅗ'); // '토'
+```
+
+## 사용해보기
+
+<br />
+
+<Sandpack>
+
+```ts index.ts
+import { combineCharacter } from 'es-hangul';
+
+console.log(combineCharacter('ㄱ', 'ㅏ', 'ㅂㅅ'));
+console.log(combineCharacter('ㅌ', 'ㅗ'));
+```
+
+</Sandpack>
