+++ src/combineCharacter/combineCharacter.spec.ts
@@ -1,4 +1,4 @@
-import { combineCharacter, combineVowels } from './combineCharacter';
+import { combineCharacter } from './combineCharacter';
 
 describe('combineCharacter', () => {
   it('종성으로 겹받침으로 구성될 수 있는 문자 두 개를 받으면 겹받침을 생성한다. (ㄱ, ㅏ, ㅂㅅ)', () => {
@@ -27,21 +27,5 @@ describe('combineCharacter', () => {
 
   it('온전한 한글 문자가 하나라도 입력되면 에러를 반환한다. (가, ㅏ, ㄱ)', () => {
     expect(() => combineCharacter('가', 'ㅏ', 'ㄱ')).toThrowError('Invalid hangul Characters: 가, ㅏ, ㄱ');
-  });
-});
-
-describe('combineVowels', () => {
-  it('겹모음이 될 수 있는 모음이 순서대로 입력되면 겹모음으로 합성한다.', () => {
-    expect(combineVowels('ㅗ', 'ㅏ')).toBe('ㅘ');
-    expect(combineVowels('ㅜ', 'ㅔ')).toBe('ㅞ');
-    expect(combineVowels('ㅡ', 'ㅣ')).toBe('ㅢ');
-  });
-  it('겹모음이 될 수 있는 모음이라고 해도 틀린 순서로 입력되면 Join한다.', () => {
-    expect(combineVowels('ㅏ', 'ㅗ')).toBe('ㅏㅗ');
-    expect(combineVowels('ㅣ', 'ㅡ')).toBe('ㅣㅡ');
-  });
-  it('이미 겹모음인 문자와 모음을 합성하려고 시도하면 Join한다.', () => {
-    expect(combineVowels('ㅘ', 'ㅏ')).toBe('ㅘㅏ');
-    expect(combineVowels('ㅝ', 'ㅣ')).toBe('ㅝㅣ');
   });
 });
