+++ src/combineCharacter/combineCharacter.ts
@@ -1,10 +1,4 @@
-import {
-  CHOSEONGS,
-  COMPLETE_HANGUL_START_CHARCODE,
-  DISASSEMBLED_VOWELS_BY_VOWEL,
-  JONGSEONGS,
-  JUNSEONGS,
-} from '@/_internal/constants';
+import { CHOSEONGS, COMPLETE_HANGUL_START_CHARCODE, JONGSEONGS, JUNSEONGS } from '@/_internal/constants';
 import { canBeChoseong } from '@/canBeChoseong';
 import { canBeJongseong } from '@/canBeJongseong';
 import { canBeJungseong } from '@/canBeJungseong';
@@ -46,39 +40,3 @@ export function combineCharacter(choseong: string, jungseong: string, jongseong 
 
   return String.fromCharCode(unicode);
 }
-
-/**
- * @name curriedCombineCharacter
- * @description
- * 인자로 초성, 중성, 종성을 받아 하나의 한글 문자를 반환하는 `combineCharacter` 함수의 커링된 버전입니다.
- * @example
- * const combineMiddleHangulCharacter = curriedCombineCharacter('ㄱ')
- * const combineLastHangulCharacter = combineMiddleHangulCharacter('ㅏ')
- * combineLastHangulCharacter('ㄱ') // '각'
- */
-export const curriedCombineCharacter =
-  (choseong: string) =>
-  (jungseong: string) =>
-  (jongseong = '') =>
-    combineCharacter(choseong, jungseong, jongseong);
-
-/**
- * @name combineVowels
- * @description
- * 인자로 두 개의 모음을 받아 합성하여 겹모음을 생성합니다. 만약 올바른 한글 규칙으로 합성할 수 없는 모음들이라면 단순 Join합니다.
- * ```typescript
- * combineVowels(
- *   // 첫 번째 모음
- *   vowel1: string
- *   // 두 번째 모음
- *   vowel2: string
- * ): string
- * ```
- * @example
- * combineVowels('ㅗ', 'ㅏ') // 'ㅘ'
- * combineVowels('ㅗ', 'ㅐ') // 'ㅙ'
- * combineVowels('ㅗ', 'ㅛ') // 'ㅗㅛ'
- */
-export const combineVowels = (vowel1: string, vowel2: string) =>
-  Object.entries(DISASSEMBLED_VOWELS_BY_VOWEL).find(([, value]) => value === `${vowel1}${vowel2}`)?.[0] ??
-  `${vowel1}${vowel2}`;
