+++ src/index.ts
@@ -3,7 +3,8 @@ export { assemble } from './assemble';
 export { canBeChoseong } from './canBeChoseong';
 export { canBeJongseong } from './canBeJongseong';
 export { canBeJungseong } from './canBeJungseong';
-export { combineCharacter, combineVowels } from './combineCharacter';
+export { combineCharacter } from './combineCharacter';
+export { combineVowels } from './combineVowels';
 export { convertQwertyToAlphabet, convertQwertyToHangul } from './convertQwertyToAlphabet';
 export { days } from './days';
 export { disassemble, disassembleToGroups } from './disassemble';
